Show duplicate agent_id warnings in TeamList instead of BotList
